+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## HEAD
+* Add possiblility to add iframes within modal nicely
 * Fix height of gallery if caption and footer are present
 * Make left and right navigation buttons in gallery smaller
 * Prevent scrolling of body when modal is open